requirements.txt
python_heatclient==1.9.0
keystoneauth1==2.20.0
pycrypto==2.6.1
python_keystoneclient==3.13.0
python_novaclient==9.1.1
PyYAML==3.12
